implement transmit controller

+++ TX/sender/include/udp_socket.hpp
@@ -18,8 +18,8 @@ public:
     UDPSocket(UDPSocket&& other) noexcept;
     UDPSocket& operator=(UDPSocket&& other) noexcept;
 
-    ssize_t send(const uint32_t* data, size_t length);
-    ssize_t send(const std::vector<uint32_t>& buffer);
+    ssize_t send(const uint8_t* data, size_t length);
+    ssize_t send(const std::vector<uint8_t>& buffer);
 
     bool is_valid() const;
